+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
-## [2.32.5] - 2022-09-19
+### Changed
 
 - Profile - Groups tabs warns user of unsaved changes
+
+## [2.32.5] - 2022-09-19
 
 ### Changed